Check for reachable hosts
